Repacking of the ProxyServiceImpl
