+++ hazelcast/src/main/java/com/hazelcast/spi/impl/NodeEngineImpl.java
@@ -34,12 +34,13 @@ import com.hazelcast.spi.ExecutionService;
 import com.hazelcast.spi.NodeEngine;
 import com.hazelcast.spi.Operation;
 import com.hazelcast.spi.PostJoinAwareService;
-import com.hazelcast.spi.ProxyService;
 import com.hazelcast.spi.ServiceInfo;
 import com.hazelcast.spi.SharedService;
 import com.hazelcast.spi.annotation.PrivateApi;
 import com.hazelcast.internal.storage.DataRef;
 import com.hazelcast.internal.storage.Storage;
+import com.hazelcast.spi.impl.proxyservice.InternalProxyService;
+import com.hazelcast.spi.impl.proxyservice.impl.ProxyServiceImpl;
 import com.hazelcast.spi.impl.waitnotifyservice.InternalWaitNotifyService;
 import com.hazelcast.spi.impl.waitnotifyservice.impl.WaitNotifyServiceImpl;
 import com.hazelcast.spi.impl.eventservice.InternalEventService;
@@ -167,7 +168,7 @@ public class NodeEngineImpl implements NodeEngine {
     }
 
     @Override
-    public ProxyService getProxyService() {
+    public InternalProxyService getProxyService() {
         return proxyService;
     }
