+++ hazelcast/src/main/java/com/hazelcast/spi/impl/ServiceManager.java
@@ -47,6 +47,7 @@ import com.hazelcast.spi.ManagedService;
 import com.hazelcast.spi.NodeEngine;
 import com.hazelcast.spi.ServiceInfo;
 import com.hazelcast.spi.annotation.PrivateApi;
+import com.hazelcast.spi.impl.proxyservice.impl.ProxyServiceImpl;
 import com.hazelcast.topic.impl.TopicService;
 import com.hazelcast.transaction.impl.TransactionManagerServiceImpl;
