+++ hazelcast/src/main/java/com/hazelcast/spi/impl/proxyservice/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Contains the API for the {@link com.hazelcast.spi.impl.proxyservice.InternalProxyService}.
+ */
+package com.hazelcast.spi.impl.proxyservice;
